UPDATE ::: Fixing null dereferencing issues
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@38501 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 487b7944c46c524a82c70a648cb355f87d3c5877

+++ src/prerna/sablecc2/reactor/app/upload/r/RReplaceDatabaseCsvUploadReactor.java
@@ -123,7 +123,7 @@ public class RReplaceDatabaseCsvUploadReactor extends AbstractReactor {
 			//			UploadUtilities.updateMetadata(this.engine.getEngineId());
 			logger.info("Complete");
 		} catch (Exception e) {
-			e.printStackTrace();
+			logger.error("StackTrace: ", e);
 			error = true;
 			if (e instanceof SemossPixelException) {
 				throw (SemossPixelException) e;
@@ -137,18 +137,20 @@ public class RReplaceDatabaseCsvUploadReactor extends AbstractReactor {
 			if(error) {
 				// need to revert
 				// delete the existing file
-				if(rEngineOldFile.exists()) {
+				if(rEngineOldFile != null && rEngineOldFile.exists()) {
 					rEngineOldFile.delete();
 				}
 				// replace it
 				try {
 					Files.copy(rEngineOldFileCopy, rEngineOldFile);
 				} catch (IOException e) {
-					e.printStackTrace();
+					logger.error("StackTrace: ", e);
 				}
 			} else {
 				// delete the copy
-				rEngineOldFileCopy.delete();
+				if (rEngineOldFileCopy != null) {
+					rEngineOldFileCopy.delete();
+				}
 			}
 		}
